refactor GUI layout into module
